+++ src/from-promise.ts
@@ -1,5 +1,5 @@
 import {IObservableValue, observable, action} from "mobx";
-import {IDENTITY, deprecated} from "./utils";
+import {IDENTITY, deprecated, invariant} from "./utils";
 
 export type PromiseState = "pending" | "fulfilled" | "rejected";
 
@@ -7,38 +7,43 @@ export const PENDING = "pending";
 export const FULFILLED = "fulfilled";
 export const REJECTED = "rejected";
 
-export interface IPromiseBasedObservable<T> {
-    readonly value: T;
-    readonly state: PromiseState;
-    readonly reason: any;
+export type IBasePromiseBasedObservable<T> = {
     readonly promise: PromiseLike<T>;
     case<U>(handlers: {pending?: () => U, fulfilled?: (t: T) => U, rejected?: (e: any) => U}): U;
 }
 
-class PromiseBasedObservable<T> implements IPromiseBasedObservable<T> {
+export type IPendingPromise = {
+    readonly state: "pending";
+    readonly reason: any;
+}
+
+export type IResolvedPromise<T> = {
+    readonly state: "fulfilled";
+    readonly value: T;
+}
+
+export type IRejectedPromise = {
+    readonly state: "rejected";
+    readonly value: any;
+}
+
+export type IPromiseBasedObservable<T> = IBasePromiseBasedObservable<T> & (IPendingPromise | IResolvedPromise<T> | IRejectedPromise)
+
+class PromiseBasedObservable<T> {
     @observable.ref value: T = undefined;
     @observable.ref state: PromiseState = PENDING;
-    @observable.ref _reason: any = undefined;
 
-    constructor(public promise: PromiseLike<T>, initialValue: T = undefined) {
-        this.value = initialValue;
+    constructor(public promise: PromiseLike<T>) {
         promise.then(
             action("observableFromPromise-resolve", (value: T) => {
                 this.value = value;
                 this.state = FULFILLED;
             }),
             action("observableFromPromise-reject", (reason: any) => {
-                this._reason = reason;
                 this.value = reason;
                 this.state = REJECTED;
             })
         );
-    }
-
-    get reason(): any {
-        // TODO: kill in next major
-        deprecated("In `fromPromise`: `.reason` is deprecated, use `.value` instead");
-        return this._reason;
     }
 
     public case<U>(handlers: {pending?: () => U, fulfilled?: (t: T) => U, rejected?: (e: any) => U}): U {
@@ -95,8 +100,10 @@ class PromiseBasedObservable<T> implements IPromiseBasedObservable<T> {
  * @param {T} [initialValue=undefined] Optional predefined initial value
  * @returns {IPromiseBasedObservable<T>}
  */
-export function fromPromise<T>(promise: PromiseLike<T>, initialValue: T = undefined): IPromiseBasedObservable<T> {
-    return new PromiseBasedObservable(promise, initialValue);
+export function fromPromise<T>(promise: PromiseLike<T>): IPromiseBasedObservable<T> {
+    invariant(arguments.length === 1, "fromPromise expects exactly one argument");
+    invariant(typeof promise === "object" && promise && typeof promise.then === "function", "Please pass a promise to fromPromise");
+    return new PromiseBasedObservable(promise) as any;
 }
 
 /**